Added more exercises and purged the other databases
